add:hook_productindex
